// src/services/api/dentistApi.ts
export interface NewDentistDetail {
  firstName: string;
  lastName: string;
  email: string;
  userName: string;
  password: string;
  isDentist: boolean;
  dob: string;
  phoneNumber: string;
  sex: boolean;
  status: number;
  description: string;
  licenseNumber: string;
  yearOfExperience: number;
  clinicDentalId: string;
}